Include preliminary info on personal projects
with their respective repository and application links.
Projects added so far are the LSTM Stock Dashboard and
JavaFX 8 Minesweeper. I am unable to embed the latter
into a webpage as a web app due to security conflicts,
so for now the executable jar file repo will suffice.
Also, I plan on adding brief descriptions of these projects
(along with sample screenshots?) later on, so for now
[Description] is present as a placeholder.